Fix bugs in comparing Route53 resources against AWS responses
* Use .all! instead of .all, to really get beyond the first batch.
* Use .chomp('.') and .to_s where necessary to compare values correctly.
* Add a check for r.type == resource[:type], to avoid destroying innocent records.
* Add mk_resource_methods, because Puppet was throwing exceptions otherwise.

+++ lib/puppet/provider/dns_record/route53.rb
@@ -38,10 +38,12 @@ Puppet::Type.type(:dns_record).provide(:route53) do
 
   desc "Manage AWS Route 53 records."
 
+  mk_resource_methods
+
   def create
     @zone = route53.zones.get(resource[:zone_id])
-    @zone.records.all.each do |r|
-      if r.name == resource[:name]
+    @zone.records.all!.each do |r|
+      if r.name.chomp('.') == resource[:name] && r.type == resource[:type]
         Puppet.debug("Route53: Cannot modify records, must remove #{resource[:name]}.")
         r.destroy
       end
@@ -64,8 +66,8 @@ Puppet::Type.type(:dns_record).provide(:route53) do
     @username, @password = resource[:username], resource[:password]
     resource[:content] = resource[:content].is_a?(Array) ? resource[:content] : resource[:content].to_a
     @zone = route53.zones.get(resource[:zone_id])
-    records = @zone.records.all
-    if records.detect { |r| r.name == resource[:name] and r.value == resource[:content] and r.ttl == resource[:ttl] }
+    records = @zone.records.all!
+    if records.detect { |r| r.name.chomp('.') == resource[:name] and r.value == resource[:content] and r.ttl == resource[:ttl].to_s }
       return true
     else
       return false
@@ -74,8 +76,8 @@ Puppet::Type.type(:dns_record).provide(:route53) do
 
   def destroy
     @zone = route53.zones.get(resource[:zone_id])
-    @zone.records.all.each do |r|
-      if ( r.name == resource[:name] ) and ( r.type == resource[:type] )
+    @zone.records.all!.each do |r|
+      if ( r.name.chomp('.') == resource[:name] ) and ( r.type == resource[:type] )
         r.destroy
         Puppet.info("Route53: destroyed #{resource[:type]} record for #{resource[:name]}.#{resource[:domain]}")
       end
